third_party/WebKit/LayoutTests/platform/win/http/tests/security/contentSecurityPolicy/worker-src/shared-none-expected.txt

CONSOLE ERROR: line 110: Refused to create a worker from 'http://127.0.0.1:8000/security/contentSecurityPolicy/resources/ping.js' because it violates the following Content Security Policy directive: "worker-src 'none'".

CONSOLE ERROR: line 110: Refused to create a worker from 'blob:http://127.0.0.1:8000/51e84c4e-5f52-4376-9ca7-25b523d4fce0' because it violates the following Content Security Policy directive: "worker-src 'none'".

This is a testharness.js-based test.
PASS Same-origin shared worker blocked by 'none'. 
PASS blob: shared worker blocked by 'none'. 
Harness: the test ran to completion.
